[#13] - (refactor) 일부 변경사항 롤백
