+++ app/api/news/news_service.py
