+++ app/api/stock/stock_router.py
@@ -1,5 +1,5 @@
 from fastapi import APIRouter, Query, HTTPException
-from app.legacy.stock.stock_service import StockService
+from app.api.stock.stock_service import StockService
 
 router = APIRouter()
